Minor clean-up
