+++ VirtualRyan.Server/ChatRequest.cs
@@ -1,0 +1,12 @@
+namespace VirtualRyan.Server.Controllers
+{
+	public partial class ChatController
+	{
+		public class ChatRequest
+		{
+			public string? Question { get; set; }
+
+			public string? ConversationId { get; set; }
+		}
+	}
+}
